Fix layer-3-deps violation: move hourly snapshot retention to contracts
packages/execution-engine/src/workers/reaper-reclaim.ts imported
hourlySnapshotRetention/pruneHourlySnapshots from packages/app, which
Dependency Cruise's layer-3-deps rule forbids (execution-engine can only
depend on layers 0-2). Move the two pure functions to @invoker/contracts,
re-exported from packages/app/src/delete-all-snapshot.ts for existing callers.
