+++ packages/app/src/delete-all-snapshot.ts
@@ -1,8 +1,12 @@
-import { copyFileSync, existsSync, mkdirSync, readdirSync, rmSync } from 'node:fs';
+import { copyFileSync, existsSync, mkdirSync } from 'node:fs';
 import * as path from 'node:path';
-import { resolveInvokerHomeRoot } from '@invoker/contracts';
+import {
+  resolveInvokerHomeRoot,
+  hourlySnapshotRetention,
+  pruneHourlySnapshots,
+} from '@invoker/contracts';
 
-export { resolveInvokerHomeRoot };
+export { resolveInvokerHomeRoot, hourlySnapshotRetention, pruneHourlySnapshots };
 
 /**
  * Caller-supplied backup implementation, expected to write a fully
@@ -66,68 +70,6 @@ export async function createDeleteAllSnapshot(
   backup?: SnapshotBackupFn,
 ): Promise<string | null> {
   return createDbSnapshot('before-delete-all', invokerHomeRoot, backup);
-}
-
-const DEFAULT_HOURLY_SNAPSHOT_RETENTION = 48;
-const HOURLY_SNAPSHOT_PREFIX = 'invoker.db.hourly-auto-';
-
-export function hourlySnapshotRetention(): number {
-  const raw = process.env.INVOKER_HOURLY_BACKUP_RETENTION;
-  // Treat empty/blank as unset: Number('') and Number('   ') are 0, which would
-  // otherwise pass the >= 0 check and silently disable pruning (reintroducing the
-  // unbounded growth this guards against). `export VAR=` should fall back, not disable.
-  if (raw === undefined || raw.trim() === '') return DEFAULT_HOURLY_SNAPSHOT_RETENTION;
-  const parsed = Number(raw);
-  return Number.isFinite(parsed) && parsed >= 0
-    ? Math.floor(parsed)
-    : DEFAULT_HOURLY_SNAPSHOT_RETENTION;
-}
-
-/**
- * Delete the oldest `hourly-auto` snapshots (and any legacy `-wal`/`-shm`
- * sidecars left over from the pre-fix raw-copy era) so at most `retain`
- * remain. Without this the hourly backup grows without bound — a single
- * host accumulated 1,554 snapshots (~363 GB). `retain <= 0` disables
- * pruning. Only `hourly-auto` snapshots are pruned; manual and
- * pre-delete-all snapshots are left untouched. Returns the number of
- * snapshots removed.
- */
-export function pruneHourlySnapshots(backupDir: string, retain: number): number {
-  if (retain <= 0) return 0;
-  let entries: string[];
-  try {
-    entries = readdirSync(backupDir);
-  } catch {
-    return 0;
-  }
-  // Base snapshot files only; the timestamp suffix (YYYYMMDD-HHMMSS-mmmZ) sorts
-  // chronologically, so the oldest snapshots come first.
-  const snapshots = entries
-    .filter(
-      (name) =>
-        name.startsWith(HOURLY_SNAPSHOT_PREFIX) &&
-        !name.endsWith('-wal') &&
-        !name.endsWith('-shm'),
-    )
-    .sort();
-  const excess = snapshots.length - retain;
-  if (excess <= 0) return 0;
-  let removed = 0;
-  for (const name of snapshots.slice(0, excess)) {
-    for (const suffix of ['', '-wal', '-shm']) {
-      try {
-        rmSync(path.join(backupDir, `${name}${suffix}`), { force: true });
-      } catch (err) {
-        console.warn(
-          `[delete-all-snapshot] failed to prune snapshot file ${name}${suffix}: ${
-            err instanceof Error ? err.message : String(err)
-          }`,
-        );
-      }
-    }
-    removed += 1;
-  }
-  return removed;
 }
 
 /**
