+++ packages/execution-engine/src/workers/reaper-reclaim.ts
@@ -11,9 +11,8 @@ import {
 import { homedir } from 'node:os';
 import { join } from 'node:path';
 
-import type { Logger } from '@invoker/contracts';
+import { hourlySnapshotRetention, pruneHourlySnapshots, type Logger } from '@invoker/contracts';
 
-import { hourlySnapshotRetention, pruneHourlySnapshots } from '../../../app/src/delete-all-snapshot.js';
 import { buildSshConnectionArgs } from '../ssh-transport-options.js';
 import { bashNormalizeTildePath, execRemoteCapture, shellPosixSingleQuote } from '../ssh-git-exec.js';
